Factor out DataTypeParsing.
